refactor init command
